refactor: move HTTP layer into backend/src/server/http/
Pure mechanical move: httpServerBase.ts, httpserver.ts (renamed to
httpServer.ts) and auth/ now live under src/server/http/. Import paths
adjusted; re-export names in server/index.ts stay stable.
