+++ backend/src/server/http/httpServerBase.ts
@@ -5,15 +5,15 @@ import { Application, NextFunction, Request } from 'express'
 import type { ParamsDictionary } from 'express-serve-static-core'
 import type { ParsedQs } from 'qs'
 import express from 'express'
-import { Config } from './config.js'
-import { HttpErrorsEnum } from '../shared/specification/index.js'
+import { Config } from '../config.js'
+import { HttpErrorsEnum } from '../../shared/specification/index.js'
 import { join, basename } from 'path'
 import { parse } from 'node-html-parser'
 import * as fs from 'fs'
-import { LogLevelEnum, Logger } from '../specification/index.js'
+import { LogLevelEnum, Logger } from '../../specification/index.js'
 
-import { apiUri } from '../shared/server/index.js'
-import { ConfigPersistence } from './persistence/configPersistence.js'
+import { apiUri } from '../../shared/server/index.js'
+import { ConfigPersistence } from '../persistence/configPersistence.js'
 import { createAuthMiddleware } from './auth/authMiddleware.js'
 import { initOidc, registerOidcRoutes, setupSession, type OidcConfig } from './auth/oidc.js'
