+++ backend/src/server/index.ts
@@ -1,9 +1,8 @@
 export * from './bus.js'
 export * from './config.js'
 export * from './configbus.js'
-export * from './httpserver.js'
-export * from './httpServerBase.js'
-export * from './index.js'
+export * from './http/httpServer.js'
+export * from './http/httpServerBase.js'
 export * from './modbus.js'
 export * from './modbusAPI.js'
 export * from './modbusRTUprocessor.js'
